refactor(llm-agents): extract the shared setAgentMaxIterations helper (#1380)
`setMaxIterations()` was copy-pasted into agent-max-iterations.spec.ts and
agent-multi-tool-selection.spec.ts, both driving the same four handles in the
same order. #1378 left the duplication deliberately rather than pull a second
@stable spec into its blast radius; this is that follow-up.

The two details the extraction had to keep, both in the helper now:

- the dev49 note explaining WHY the field is reached through the inspector
  (it replaced the Controls dialog / edit-button-modal), which is the reason
  the handle sequence looks the way it does;
- the toHaveValue read-back that only the agent-multi-tool-selection copy had.
  What it buys is not the same on both sides, and the header says so rather
  than generalising: in agent-multi-tool-selection a no-op fill leaves the
  default 15 and re-opens #1378 on a run that still looks green; in
  agent-max-iterations Test 1 goes red either way, so the read-back buys
  attribution — it stops at the fill instead of timing out 30 s later on the
  missing limit message.

Knowledge stays where it belongs: the #1378 measurement block (pass rates, the
recursion_limit asymmetry, the #14489 upstream cap) documents the VALUE 8, so
it moved to MAX_ITERATIONS_SEQUENCE; the helper documents only the mechanism.

addAgentFieldsToBody in agent-config-persistence.spec.ts drives the same four
handles and is deliberately left alone. Merging the two functions is wrong on
its merits — that spec batches several adds and must let the add-autosave
settle before writing any value. Extracting the add loop and layering this
helper on top would work, and was declined for scope: it pulls a third @stable
spec into this PR.
